# Performance bugfix: >>100% performance improvements in Detection project