--- a/Detection/src/Services/PlatformService.cs
+++ b/Detection/src/Services/PlatformService.cs
@@ -11,8 +11,8 @@ namespace Wangkanai.Detection.Services;
 public sealed class PlatformService : IPlatformService
 {
 	private readonly IUserAgentService _userAgentService;
-	private          Platform?         _name;
 
+	private Platform?  _name;
 	private Processor? _processor;
 	private Version?   _version;
 
@@ -131,26 +131,27 @@ public sealed class PlatformService : IPlatformService
 	private static string AgentSourceStart(string agent, string prefix)
 		=> _osStartRegex.RegexMatch(agent)
 		                .Captures
-		                .FirstOrDefault()?
+		                .FirstOrDefault()
+		                ?
 		                .Value
 		                .RemoveAll(" ", "(", ")")
 		                .Split(';')
 		                .FirstOrDefault(x => x.StartsWith(prefix, StringComparison.Ordinal));
 
 
-	private static readonly string[]   X86DeviceList    = { "i86", "i686", Processor.x86.ToString() };
-	private static readonly IPrefixTrie X86DeviceTrie   = X86DeviceList.BuildSearchTrie();
-	private static readonly string[]   X64DeviceList    = { "x86_64", "wow64", "win64", Processor.x64.ToLowerString() };
-	private static readonly IPrefixTrie X64DeviceTrie   = X64DeviceList.BuildSearchTrie();
-	private static readonly string[]   IosDeviceList    = { "iphone", "ipod", Platform.iOS.ToLowerString() };
-	private static readonly IPrefixTrie IosDeviceTrie   = IosDeviceList.BuildSearchTrie();
-	private static readonly string[]   IPadosDeviceList = { "ipad", Platform.iPadOS.ToLowerString() };
+	private static readonly string[]    X86DeviceList    = { "i86", "i686", Processor.x86.ToString() };
+	private static readonly IPrefixTrie X86DeviceTrie    = X86DeviceList.BuildSearchTrie();
+	private static readonly string[]    X64DeviceList    = { "x86_64", "wow64", "win64", Processor.x64.ToLowerString() };
+	private static readonly IPrefixTrie X64DeviceTrie    = X64DeviceList.BuildSearchTrie();
+	private static readonly string[]    IosDeviceList    = { "iphone", "ipod", Platform.iOS.ToLowerString() };
+	private static readonly IPrefixTrie IosDeviceTrie    = IosDeviceList.BuildSearchTrie();
+	private static readonly string[]    IPadosDeviceList = { "ipad", Platform.iPadOS.ToLowerString() };
 
 	private static readonly IPrefixTrie PadosDeviceTrie = IPadosDeviceList.BuildSearchTrie();
-	private static readonly string[]   ChromeOSList      = { "cros" };
-	private static readonly IPrefixTrie ChromeOsTrie     = ChromeOSList.BuildSearchTrie();
-	private static readonly string[]   AppleWebKitList   = { "applewebkit", "webkit" };
-	private static readonly IPrefixTrie AppleWebKitTrie  = AppleWebKitList.BuildSearchTrie();
+	private static readonly string[]    ChromeOSList    = { "cros" };
+	private static readonly IPrefixTrie ChromeOsTrie    = ChromeOSList.BuildSearchTrie();
+	private static readonly string[]    AppleWebKitList = { "applewebkit", "webkit" };
+	private static readonly IPrefixTrie AppleWebKitTrie = AppleWebKitList.BuildSearchTrie();
 
 	#endregion
 }