--- a/System/src/Extensions/EnumExtensions.cs
+++ b/System/src/Extensions/EnumExtensions.cs
@@ -94,7 +94,7 @@ public static class EnumExtensions
 	/// <typeparam name="T">The type of the enumeration.</typeparam>
 	/// <param name="value">The enumeration value to retrieve the flags from.</param>
 	/// <returns>An enumerable collection of the individual flags of the enumeration value.</returns>
-	public static IEnumerable<T> GetFlags<T>(this T value)
+	public static IReadOnlySet<T> GetFlags<T>(this T value)
 		where T : Enum
 		=> EnumValues<T>.GetFlags(value);