--- a/System/src/Extensions/EnumValues.cs
+++ b/System/src/Extensions/EnumValues.cs
@@ -1,4 +1,6 @@
-// Copyright (c) 2014-2024 Sarin Na Wangkanai, All Rights Reserved.Apache License, Version 2.0
+// Copyright (c) 2014-2024 Sarin Na Wangkanai and Aliaksandr Kukrash, All Rights Reserved.Apache License, Version 2.0
+
+using System.Collections.Concurrent;
 
 namespace Wangkanai.Extensions;
 
@@ -12,37 +14,71 @@ namespace Wangkanai.Extensions;
 [DebuggerStepThrough]
 public static class EnumValues<T> where T : Enum
 {
-	private static readonly T[]                Values    = (T[])Enum.GetValues(typeof(T));
-	private static readonly Dictionary<T, T[]> EnumFlags = Values.ToDictionary(v => v, v => Values.Where(item => v.HasFlag(item)).ToArray());
+	private static readonly ConcurrentDictionary<T, IReadOnlySet<T>> MultiValueFlagsCache = new();
+	private static readonly ConcurrentDictionary<T, EnumValueName>   MultiValueCache      = new();
 
-	private static readonly Dictionary<string, T>    ValuesOriginal   = Values.ToDictionary(value => value.ToString(), value => value, StringComparer.Ordinal);
-	private static readonly Dictionary<string, T>    ValuesIgnoreCase = Values.ToDictionary(value => value.ToString(), value => value, StringComparer.OrdinalIgnoreCase);
+	private static readonly T[]                Values    = (T[]) Enum.GetValues(typeof(T));
+	private static readonly Dictionary<T, IReadOnlySet<T>> EnumFlags = Values.ToDictionary(v => v, v => (IReadOnlySet<T>)Values.Where(item => v.HasFlag(item)).ToHashSet());
 
-	private static readonly Dictionary<T, EnumValueName> NamesOriginal
-		= Values.ToDictionary(value => value, value => new EnumValueName
-		                                               {
-			                                               Name          = value.ToString(),
-			                                               NameWithFlags = GetNameWithFlags(value)
-		                                               });
-	private static readonly Dictionary<T, EnumValueName> NamesLower
-		= Values.ToDictionary(value => value, value => new EnumValueName
-		                                               {
-			                                               Name          = value.ToString().ToLowerInvariant(),
-			                                               NameWithFlags = GetNameWithFlags(value).ToLowerInvariant()
-		                                               });
+	private static readonly Dictionary<string, T> ValuesOriginal   = Values.ToDictionary(value => value.ToString(), value => value, StringComparer.Ordinal);
+	private static readonly Dictionary<string, T> ValuesIgnoreCase = Values.ToDictionary(value => value.ToString(), value => value, StringComparer.OrdinalIgnoreCase);
 
-	private static readonly Dictionary<T, EnumValueName> NamesUpper
-		= Values.ToDictionary(value => value, value => new EnumValueName
-		                                               {
-			                                               Name          = value.ToString().ToUpperInvariant(),
-			                                               NameWithFlags = GetNameWithFlags(value).ToUpperInvariant()
-		                                               });
+	private static readonly Dictionary<T, string>                  PlainNames      = Values.ToDictionary(value => value, value => value.ToString());
 
 	private static string GetNameWithFlags(T value)
 	{
 		return value.GetFlags().Any()
 			       ? string.Join(',', value.GetFlags().Select(x => x.ToString()))
 			       : value.ToString();
+	}
+
+	private enum ValueKind
+	{
+		Normal = 0,
+		Lower  = 1,
+		Upper  = 2,
+	}
+
+	private static string GetNameWithFlagsCached(T value, bool returnAllFlags, ValueKind kind)
+	{
+		var names = MultiValueCache.GetOrAdd(value,
+			v => new EnumValueName
+			     {
+				     Name               = v.ToString(),
+				     NameWithFlags      = GetNameWithFlags(v),
+				     NameLower          = v.ToString().ToLowerInvariant(),
+				     NameLowerWithFlags = GetNameWithFlags(v).ToLowerInvariant(),
+				     NameUpper          = v.ToString().ToUpperInvariant(),
+				     NameUpperWithFlags = GetNameWithFlags(v).ToUpperInvariant()
+			     });
+		if (returnAllFlags)
+		{
+			switch (kind)
+			{
+				case ValueKind.Normal:
+					return names.NameWithFlags;
+				case ValueKind.Lower:
+					return names.NameLowerWithFlags;
+				case ValueKind.Upper:
+					return names.NameUpperWithFlags;
+				default:
+					throw new ArgumentOutOfRangeException(nameof(kind), kind, null);
+			}
+		}
+		else
+		{
+			switch (kind)
+			{
+				case ValueKind.Normal:
+					return names.Name;
+				case ValueKind.Lower:
+					return names.NameLower;
+				case ValueKind.Upper:
+					return names.NameUpper;
+				default:
+					throw new ArgumentOutOfRangeException(nameof(kind), kind, null);
+			}
+		}
 	}
 
 	/// <summary>
@@ -53,18 +89,23 @@ public static class EnumValues<T> where T : Enum
 	public static T[] GetValues()
 		=> Values;
 
-	public static T[] GetFlags(T value)
+	public static IReadOnlySet<T> GetFlags(T value)
 	{
-		return EnumFlags[value];
+		if (EnumFlags.TryGetValue(value, out var flags))
+		{
+			return flags;
+		}
+
+		return MultiValueFlagsCache.GetOrAdd(value, v => Values.Where(item => v.HasFlag(item)).ToHashSet());
 	}
 
 	/// <summary>
-	/// Retrieves a dictionary of all names corresponding to the values of the specified enum type.
+	/// Retrieves all names corresponding to the values of the specified enum type.
 	/// </summary>
 	/// <typeparam name="T">The type of the enum.</typeparam>
-	/// <returns>A dictionary of all names corresponding to the values of the specified enum type.</returns>
-	public static Dictionary<T, EnumValueName> GetNames()
-		=> NamesOriginal;
+	/// <returns>All names corresponding to the values of the specified enum type.</returns>
+	public static IEnumerable<string> GetNames()
+		=> Values.Select(v => GetNameWithFlagsCached(v, false, ValueKind.Normal));
 
 	/// <summary>
 	/// Tries to retrieve the single name of the specified enumeration value.
@@ -73,18 +114,7 @@ public static class EnumValues<T> where T : Enum
 	/// <param name="value">The enumeration value to get the name of.</param>
 	/// <param name="result">When this method returns, contains the single name associated with the specified enumeration value, if the retrieval succeeds; otherwise, an empty string. This parameter is passed uninitialized.</param>
 	/// <returns><see langword="true"/> if the single name of the specified enumeration value is successfully retrieved; otherwise <see langword="false"/>.</returns>
-	public static bool TryGetSingleName(T value, [MaybeNullWhen(false)] out string result)
-	{
-		var exist = NamesOriginal.TryGetValue(value, out var names);
-		if (exist)
-		{
-			result = names.Name;
-			return true;
-		}
-
-		result = null;
-		return false;
-	}
+	public static bool TryGetSingleName(T value, [MaybeNullWhen(false)] out string result) => PlainNames.TryGetValue(value, out result);
 
 	/// <summary>
 	/// Retrieves the original name(s) of the specified enum value.
@@ -92,7 +122,8 @@ public static class EnumValues<T> where T : Enum
 	/// <typeparam name="T">The type of the enum.</typeparam>
 	/// <param name="value">The enum value to retrieve the original name(s) for.</param>
 	/// <returns>The original name(s) of the specified enum value.</returns>
-	public static string GetNameOriginal(T value) => NamesOriginal[value].NameWithFlags;
+	public static string GetNameOriginal(T value) =>
+		GetNameWithFlagsCached(value, true, ValueKind.Normal);
 
 	/// <summary>
 	/// Retrieves the lowercased name of the specified enum value.
@@ -101,7 +132,7 @@ public static class EnumValues<T> where T : Enum
 	/// <param name="value">The enum value.</param>
 	/// <returns>The lowercased name of the enum value.</returns>
 	public static string GetNameLower(T value)
-		=> NamesLower[value].NameWithFlags;
+		=> GetNameWithFlagsCached(value, true, ValueKind.Lower);
 
 	/// <summary>
 	/// Retrieves the uppercased name of the specified enum value.
@@ -110,23 +141,48 @@ public static class EnumValues<T> where T : Enum
 	/// <param name="value">The enum value.</param>
 	/// <returns>The uppercased name of the specified enum value.</returns>
 	public static string GetNameUpper(T value)
-		=> NamesUpper[value].NameWithFlags;
+		=> GetNameWithFlagsCached(value, true, ValueKind.Upper);
 
+	/// <summary>
+	/// Returns corresponding enum value based on its string representation
+	/// </summary>
+	/// <param name="valueName"></param>
+	/// <param name="ignoreCase"></param>
+	/// <returns>Enum value</returns>
+	/// <exception cref="ArgumentException"></exception>
 	public static T Parse(string valueName, bool ignoreCase)
 	{
 		if (ignoreCase)
 		{
-			return ValuesIgnoreCase[valueName];
+			if (ValuesIgnoreCase.TryGetValue(valueName, out var result))
+			{
+				return result;
+			}
+			else
+			{
+				throw new ArgumentException($"The value '{valueName}' is not recognized.");
+			}
 		}
 		else
 		{
-			return ValuesOriginal[valueName];
+			if (ValuesOriginal.TryGetValue(valueName, out var result))
+			{
+				return result;
+			}
+			else
+			{
+				throw new ArgumentException($"The value '{valueName}' is not recognized.");
+			}
 		}
 	}
-}
 
-public readonly struct EnumValueName
-{
-	public string Name          { get; init; }
-	public string NameWithFlags { get; init; }
+	private readonly struct EnumValueName
+	{
+		public required string Name               { get; init; }
+		public required string NameWithFlags      { get; init; }
+		public required string NameLower          { get; init; }
+		public required string NameLowerWithFlags { get; init; }
+		public required string NameUpper          { get; init; }
+		public required string NameUpperWithFlags { get; init; }
+	}
 }