--- a/System/src/Reflection/AssemblyVersion.cs
+++ b/System/src/Reflection/AssemblyVersion.cs
@@ -2,15 +2,22 @@
 
 namespace Wangkanai.Reflection;
 
+/// <summary>
+/// Provides utility methods and properties for retrieving version information of the assembly.
+/// </summary>
 public static class AssemblyVersion
 {
-	public static readonly string Version
-		= AssemblyInformationVersion<AssemblyInformationalVersionAttribute>().InformationalVersion;
-
-	private static readonly Assembly EntryAssembly
-		= Assembly.GetEntryAssembly().ThrowIfNull()!;
-
-	private static T AssemblyInformationVersion<T>()
-		where T : Attribute
+	private static T AssemblyInformationVersion<T>() where T : Attribute
 		=> EntryAssembly.GetCustomAttribute<T>().ThrowIfNull()!;
+
+	private static readonly Assembly EntryAssembly = Assembly.GetEntryAssembly().ThrowIfNull()!;
+
+	/// <summary>
+	/// Represents the version of the assembly as retrieved from the assembly's informational version attribute.
+	/// </summary>
+	/// <remarks>
+	/// The version is determined by extracting the informational version from the entry assembly's metadata,
+	/// which is specified using the <see cref="AssemblyInformationalVersionAttribute"/> attribute.
+	/// </remarks>
+	public static readonly string Version = AssemblyInformationVersion<AssemblyInformationalVersionAttribute>().InformationalVersion;
 }