… Markdown/src/Infrastructure/DynamicMarkdownEndpointSelector.cs
… System/tests/Extensions/EnumValuesTests.cs
… System/tests/Extensions/SerializationExtensionsTests.cs