Here are some substantial optimizations for the Detection project and some portions of System project.

- Be careful with property declaration syntax (=>). This fires back with with unexpected performance consequences (stage 1)
- Enum from string optimizations (stage 2)
- EnumFlags optimizations (stage 3)

I only included Detection benchmarks down below.
Here are the results for all 3 optimization stages

_Original result from here: #1025_ 
Method | Mean | Error | StdDev | Gen0 | Gen1 | Allocated
-- | -- | -- | -- | -- | -- | --
Device | 12.649 μs | 0.1200 μs | 0.1064 μs | 2.5177 | 0.0153 | 41.25 KB
Platform | 9.664 μs | 0.1038 μs | 0.0867 μs | 2.5330 | 0.0153 | 41.63 KB
Engine | 37.763 μs | 0.3858 μs | 0.3420 μs | 8.7891 | - | 144.4 KB
Browser | 69.916 μs | 0.5773 μs | 0.5400 μs | 15.5029 | - | 254.16 KB

_Stage 1_
Method | Mean | Error | StdDev | Gen0 | Allocated
-- | -- | -- | -- | -- | --
Device | 4.879 μs | 0.0955 μs | 0.1570 μs | 0.9613 | 15.75 KB
Platform | 2.297 μs | 0.0419 μs | 0.0372 μs | 0.9842 | 16.13 KB
Engine | 7.589 μs | 0.0816 μs | 0.0763 μs | 2.7161 | 44.38 KB
Browser | 15.784 μs | 0.3051 μs | 0.2854 μs | 4.6082 | 75.38 KB

_Stage 2_
Method | Mean | Error | StdDev | Gen0 | Allocated
-- | -- | -- | -- | -- | --
Device | 3.708 μs | 0.0727 μs | 0.1174 μs | 0.6485 | 10.63 KB
Platform | 1.393 μs | 0.0141 μs | 0.0125 μs | 0.6733 | 11 KB
Engine | 3.602 μs | 0.0364 μs | 0.0304 μs | 1.4610 | 23.88 KB
Browser | 8.269 μs | 0.1340 μs | 0.1187 μs | 2.4109 | 39.5 KB

_Stage 3 (final result)_
Method | Mean | Error | StdDev | Gen0 | Allocated
-- | -- | -- | -- | -- | --
Device | 3.440 μs | 0.0541 μs | 0.0555 μs | 0.5569 | 9.13 KB
Platform | 1.301 μs | 0.0120 μs | 0.0106 μs | 0.5798 | 9.5 KB
Engine | 3.411 μs | 0.0362 μs | 0.0321 μs | 1.0910 | 17.88 KB
Browser | 7.871 μs | 0.1537 μs | 0.1645 μs | 1.7700 | 29 KB

It is worth noting how much better memory usage become, not just execution time.

_Setup_
BenchmarkDotNet v0.13.12, Windows 11 (10.0.26100.3476)
AMD Ryzen 9 9950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK 9.0.201
  [Host]     : .NET 8.0.14 (8.0.1425.11118), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  DefaultJob : .NET 8.0.14 (8.0.1425.11118), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI