@wangkanai 2025-04-03
Wow! @multiarc, I can not believe what I am seeing. You are super impressive.

I will have refactoring later this afternoon when I arrive back in Bangkok.