@wangkanai 2025-04-03
I follow your process and refactor fine-tuning the code also.

Baseline from #1039 

| Method   | Mean      | Error     | StdDev    | Gen0   | Allocated |
|--------- |----------:|----------:|----------:|-------:|----------:|
| Device   |  6.714 us | 0.1031 us | 0.0965 us | 1.9226 |  15.75 KB |
| Platform |  2.810 us | 0.0440 us | 0.0390 us | 1.9722 |  16.13 KB |
| Engine   |  8.652 us | 0.0880 us | 0.0823 us | 5.4321 |  44.38 KB |
| Browser  | 16.870 us | 0.1008 us | 0.0943 us | 9.2163 |  75.38 KB |

PR final results 

| Method   | Mean     | Error     | StdDev    | Gen0   | Allocated |
|--------- |---------:|----------:|----------:|-------:|----------:|
| Device   | 5.630 us | 0.1112 us | 0.1594 us | 1.1139 |   9.13 KB |
| Platform | 1.603 us | 0.0169 us | 0.0159 us | 1.1616 |    9.5 KB |
| Engine   | 4.211 us | 0.0423 us | 0.0395 us | 2.1820 |  17.88 KB |
| Browser  | 8.928 us | 0.0742 us | 0.0694 us | 3.5400 |     29 KB |

Refactor 1

| Method   | Mean     | Error     | StdDev    | Gen0   | Allocated |
|--------- |---------:|----------:|----------:|-------:|----------:|
| Device   | 5.584 us | 0.1081 us | 0.1110 us | 1.1139 |   9.13 KB |
| Platform | 1.615 us | 0.0116 us | 0.0103 us | 1.1616 |    9.5 KB |
| Engine   | 4.265 us | 0.0426 us | 0.0399 us | 2.1820 |  17.88 KB |
| Browser  | 8.673 us | 0.0383 us | 0.0358 us | 3.5400 |     29 KB |

Refactor 2

| Method   | Mean     | Error     | StdDev    | Gen0   | Allocated |
|--------- |---------:|----------:|----------:|-------:|----------:|
| Device   | 5.481 us | 0.1090 us | 0.1020 us | 1.1139 |   9.13 KB |
| Platform | 1.601 us | 0.0158 us | 0.0148 us | 1.1616 |    9.5 KB |
| Engine   | 4.187 us | 0.0337 us | 0.0316 us | 2.1820 |  17.88 KB |
| Browser  | 8.809 us | 0.0434 us | 0.0406 us | 3.5400 |     29 KB |


**Benchmark environment**
BenchmarkDotNet v0.13.12, macOS 15.3.2 (24D81) [Darwin 24.3.0]
Apple M4 Max, 1 CPU, 16 logical and 16 physical cores
.NET SDK 9.0.202
  [Host]     : .NET 8.0.14 (8.0.1425.11118), Arm64 RyuJIT AdvSIMD
  DefaultJob : .NET 8.0.14 (8.0.1425.11118), Arm64 RyuJIT AdvSIMD