@wangkanai 2025-04-03
This pull request includes several changes across multiple files to improve code consistency, performance, and readability. The most important changes include updates to string handling, caching mechanisms, and code simplifications.

### String Handling Improvements:
* Updated `Contains` method in `UserAgentExtensions.cs` to use `StringComparison.OrdinalIgnoreCase` for better performance and consistency.
* Modified `Contains` method to simplify flag checking logic by using `GetFlags` method directly.

### Caching and Performance Enhancements:
* Converted `Crawlers` and `CrawlerTrie` properties in `CrawlerService.cs` to static readonly fields to improve performance and reduce memory usage.
* Added caching for multi-value flags and names in `EnumValues.cs` to optimize repeated access to enum values and their names.

### Code Simplifications:
* Simplified `GetFlags` method in `EnumExtensions.cs` to return `IReadOnlySet<T>` instead of `IEnumerable<T>`.
* Refactored `Parse` method in `EnumValues.cs` to use cached dictionaries for case-insensitive and case-sensitive lookups.

These changes collectively enhance the maintainability and efficiency of the codebase.